fix(groww): widen security_id u32 -> u64 so Groww 64-bit tokens form candles (#1244)
* fix(groww): widen security_id u32->u64 so Groww 64-bit tokens form candles

Root cause: the in-memory security_id guard was u32, truncating Groww's
64-bit instrument tokens before candle aggregation. Widen ParsedTick /
BufferedSeal / SerializedSeal security_id to u64 across the hot path and
storage; full id round-trips through the 120-128 spill region. Zero DB
migration (LONG column already i64). Verified: trading 1750 + app-groww 72
tests incl bit-62 regression proof; storage 716 lib tests green.


* fix(groww): widen bar_cache + quote read paths to u64 for restart-safe 64-bit ids

---------

+++ crates/app/src/main.rs
@@ -498,7 +498,7 @@ async fn main() -> Result<()> {
                 }
                 metrics::counter!("tv_tick_gap_summary_total").increment(1);
                 metrics::gauge!("tv_tick_gap_instruments_silent").set(total_silent as f64);
-                let top: Vec<(u32, &'static str, u64)> = gaps
+                let top: Vec<(u64, &'static str, u64)> = gaps
                     .iter()
                     .take(10)
                     .map(|(id, seg, gap)| (*id, seg.as_str(), *gap))
@@ -6322,7 +6322,7 @@ async fn start_dhan_lane(
                         // is NOT a SLO-excluded SID. N=10 is comfortably
                         // above 4 (the entire universe) so we always see
                         // the full picture when called.
-                        const SLO_TICK_FRESHNESS_EXCLUDED_SIDS: &[u32] = &[21]; // INDIA VIX
+                        const SLO_TICK_FRESHNESS_EXCLUDED_SIDS: &[u64] = &[21]; // INDIA VIX
                         const SLO_TICK_FRESHNESS_SCAN_TOP_N: usize = 10;
                         let tick_freshness = if !in_market {
                             1.0
